requirements.txt
Django==4.2.5
django-environ==0.9.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
requests==2.31.0
English-to-Hindi==0.0.1
Pillow==10.0.0